Move getSchemaAsString to BigQuerySplitManager
